Cast port Shouldn't ever overflow a 32 bit integer, as that would be a crazy high port that 32 bit apps couldn't handle!

+++ CK2SSHCredential.m
@@ -520,7 +520,7 @@ void freeKeychainContent(void *ptr, void *info)
                                                         "error description"),
                      [credential user],
                      host,
-                     port];
+                     (int)port];
     
     // Note a crazily empty password
     if (![password length]) opDescription = [opDescription stringByAppendingFormat:
